eleng/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.7
django-ckeditor==6.7.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-js-asset==2.1.0
django-mptt==0.15.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
flower==2.0.1
hiredis==2.3.2
humanize==4.9.0
idna==3.6
kombu==5.3.7
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.4
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
python-dotenv